GameContext has been added. Up and running
